Split off tokens into their own file
