+++ mace/src/code_editor.rs
@@ -8,7 +8,8 @@ use {
         range_set::{RangeSet, Span},
         size::Size,
         text::Text,
-        token_cache::{Keyword, LineInfos, Punctuator, TokenCache, TokenKind},
+        token::{Keyword, Punctuator, TokenKind},
+        token_cache::{LineInfos, TokenCache},
     },
     makepad_render::*,
     makepad_widget::*,
