+++ mace/src/lib.rs
@@ -17,5 +17,6 @@ pub mod size;
 pub mod splitter;
 pub mod tab_bar;
 pub mod text;
+pub mod token;
 pub mod token_cache;
 pub mod tree_logic;
